Add unit tests for TodosController functionalities
This commit introduces a comprehensive suite of unit tests for the `TodosController` in the web application. The tests cover key functionalities including adding, copying, deleting, editing, and changing the status of todo items. Each functionality is organized into dedicated test classes for improved structure and maintainability. Mocks for dependencies are initialized, and the controller context is set up to ensure independent test execution. The tests validate expected outcomes and interactions with mocked services, confirming the controller's behavior across various scenarios.
